pyproject.toml
[project]
name = "modulo-67"
version = "0.1.0"
description = ""
authors = [
    {name = "Kaio Grativol Baldo",email = "78832356+KaioGBaldo@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
